@@ -34,7 +34,7 @@
         "@ant-design/icons-vue": "^6.0.1",
         "@popjs/util": "^0.3",
         "@vant/touch-emulator": "^1.3.2",
-        "axios": "^0.21.1",
+        "axios": "^1.1.3",
         "better-mock": "^0.3.2",
         "crypto-js": "^4.0.0",
         "lodash-es": "^4.17.21",
@@ -47,18 +47,20 @@
         "@types/crypto-js": "^4.0.2",
         "@types/lodash-es": "^4.17.4",
         "@types/node": "^16.4.13",
-        "@typescript-eslint/eslint-plugin": "^4.21.0",
-        "@typescript-eslint/parser": "^4.21.0",
-        "@vitejs/plugin-vue": "^1.2.1",
+        "@typescript-eslint/eslint-plugin": "^5.43.0",
+        "@typescript-eslint/parser": "^5.43.0",
+        "@vitejs/plugin-vue": "^3.2.0",
         "@vue/compiler-sfc": "^3.0.5",
-        "eslint": "^7.24.0",
+        "consola": "^2.15.3",
+        "eslint": "^8.28.0",
         "eslint-plugin-import": "^2.22.1",
-        "eslint-plugin-vue": "^7.8.0",
+        "eslint-plugin-vue": "^9.7.0",
         "less": "^4.1.1",
-        "lint-staged": "^10.5.4",
+        "lint-staged": "^13.0.3",
         "typescript": "^4.1.3",
-        "vite": "^2.1.5",
-        "vite-plugin-style-import": "^0.9.2",
-        "vue-tsc": "^0.0.15"
+        "unplugin-vue-components": "^0.22.9",
+        "vite": "^3.2.4",
+        "vite-plugin-style-import": "^2.0.0",
+        "vue-tsc": "^1.0.9"
     }
 }
