@@ -7,10 +7,6 @@
     </div>
 </template>
 
-<script lang="ts">
-import { defineComponent } from 'vue'
+<script lang="ts" setup>
 
-export default defineComponent({
-    name: 'Misc'
-})
 </script>
