@@ -9,10 +9,5 @@
     </div>
 </template>
 
-<script lang="ts">
-import { defineComponent } from 'vue'
-
-export default defineComponent({
-    name: 'Vant'
-})
+<script lang="ts" setup>
 </script>
