@@ -5,21 +5,11 @@
     </div>
 </template>
 
-<script lang="ts">
-import { defineComponent, ref } from 'vue'
-import IconFont from '@/components/core/IconFont';
+<script lang="ts" setup>
+import { ref } from 'vue'
 
-export default defineComponent({
-    name: 'Home',
-    components: {
-        IconFont
-    },
-    setup() {
-        const count = ref<number>(0)
-        const increment = () => {
-            count.value += 1
-        }
-        return { count, increment }
-    }
-})
+const count = ref<number>(0)
+const increment = () => {
+    count.value += 1
+}
 </script>
