@@ -8,10 +8,6 @@
     </div>
 </template>
 
-<script lang="ts">
-import { defineComponent } from 'vue'
+<script lang="ts" setup>
 
-export default defineComponent({
-    name: 'Vue'
-})
 </script>
