@@ -1,36 +1,26 @@
 <template>
     <div class="vant--vuex">
         <div class="page-title">Vuex Test Page</div>
-        <p>store Root is: {{ text }}</p>
-        <p>store doubleCount is: {{ count }}</p>
+        <p>store Root is: {{ reactiveData.text }}</p>
+        <p>store doubleCount is: {{ reactiveData.count }}</p>
         <van-button type="primary" @click="double">double</van-button>
     </div>
 </template>
 
-<script lang="ts">
-import { computed, defineComponent, reactive, toRefs } from 'vue'
+<script lang="ts" setup>
+import { computed, reactive } from 'vue'
 import { useStore } from '@/store'
 
-export default defineComponent({
-    name: 'Vuex',
 
-    setup() {
-        const store = useStore()
-        const reactiveData = reactive({
-            text: computed(() => store.state.text),
-            count: computed(() => store.state.vant.count)
-        })
-
-        const double = () => {
-            store.commit('vant/DOUBLE_COUNT')
-        }
-
-        return {
-            ...toRefs(reactiveData),
-            double
-        }
-    }
+const store = useStore()
+const reactiveData = reactive({
+    text: computed(() => store.state.text),
+    count: computed(() => store.state.vant.count)
 })
+
+const double = () => {
+    store.commit('vant/DOUBLE_COUNT')
+}
 </script>
 
 <style scoped lang="less">
